+++ DifferentiationInterface/test/Back/FiniteDiff/benchmark.jl
@@ -1,11 +1,13 @@
 using Pkg
 
 using ADTypes: ADTypes
+using DataFrames: DataFrame
 using DifferentiationInterface, DifferentiationInterfaceTest
 import DifferentiationInterface as DI
 import DifferentiationInterfaceTest as DIT
 using FiniteDiff: FiniteDiff
 using Test
+import Chairmarks
 
 @testset "Benchmarking sparse" begin
     filtered_sparse_scenarios = filter(sparse_scenarios(; band_sizes = [])) do scen
@@ -21,7 +23,7 @@ using Test
         benchmark = :prepared,
         excluded = SECOND_ORDER,
         logging = LOGGING,
-    )
+    ) |> DataFrame
     @testset "Analyzing benchmark results" begin
         @testset "$(row[:scenario])" for row in eachrow(data)
             @test row[:allocs] == 0