+++ DifferentiationInterfaceTest/src/tests/type_stability.jl
@@ -1,0 +1,10 @@
+"""
+    test_jet(backend, scenario; subset, ignored_modules, function_filter)
+
+Perform the actual diagnosis of type stability for preparation and differentiatioj.
+
+!!! warning
+    Implemented in a package extension that depends on [JET.jl](https://github.com/aviatesk/JET.jl).
+    If this function fails with a `MethodError`, try `import JET` before running it again.
+"""
+function test_jet end