+++ DifferentiationInterfaceTest/test/runtests.jl
@@ -4,6 +4,8 @@ using DifferentiationInterfaceTest
 using Pkg
 using SparseConnectivityTracer
 using Test
+import JET
+import Chairmarks
 
 using DifferentiationInterfaceTest:
     default_scenarios,
@@ -27,6 +29,7 @@ LOGGING = get(ENV, "CI", "false") == "false"
     if GROUP == "Formalities" || GROUP == "All"
         @testset verbose = true "Formalities" begin
             include("formalities.jl")
+            include("benchmark.jl")
         end
         @testset verbose = true "Scenarios" begin
             include("scenario.jl")